Seems like silly errors
NOT TESTED. Edited directly on github. The function in_bounds doesn't handle the edge case of the hitbox for the line that connects the last point to the first point (now added line 56). The x bounds check (starting line 60) had was not checking the correct values in the if statement (was actually checking the y-value of the first point rather than the x-value of the second point).